Hide the state-internal stack Vec
This decouples the implementation of the internal stack from any
users.

I'm not really happy about the fact that getting a useful iterator
over stack items now requires a full Vec clone, but we cannot return a
trait object directly (unknown size), and constructing a new iterator
type is also non-trivial, as it would also have to create a snapshot
of the stack.

This is probably just a cost we have to pay, as if we return an
iterator over the stack, but the stack gets modified mid-way through
iteration, how is the iterator supposed to react to that? By creating
an explicit copy we avoid these kinds of issues.