Check for overflows where we easily can
This also includes a refactor of how we manipulate the top (end) of
the stack. Rather than popping and pushing, we try to prefer using
std::mem::replace to replace the top item in place, reducing the risk
that we have to allocate memory.

Also convert the peek-check-action pattern for fallible operations
with a .map() call, which is more concise.